requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openpyxl==3.1.2
Pillow==10.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
requests==2.31.0
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
aiohttp==3.9.3
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
frozenlist==1.5.0
multidict==6.2.0
yarl==1.19.0
propcache==0.3.1
